Refactor V1Referrals to match resource patterns
