--- src/CareTogether.Core/Resources/V1Referrals/V1ReferralStatus.cs
@@ -1,9 +1,0 @@
-namespace CareTogether.Resources.V1Referrals
-{
-    public enum V1ReferralStatus
-    {
-        Open,
-        Accepted,
-        Closed,
-    }
-}
